Difference absolute sample counts in integers, not f32
The re-anchor term converted two absolute sample counts to f32 before
subtracting them. The counts differ by at most one sample, but past 2^24
samples -- 349 seconds at 48 kHz -- an f32 ulp exceeds one, so the
difference came back as 0, +/-2, or worse instead of +/-1.

That term is the sub-sample anchor every predicted tick chain starts
from, so a dropout in any session running longer than about six minutes
put a whole-sample error, and in the worst measured case fifteen samples
-- 179 degrees of bearing -- on every coasted tick, with lock quality
reporting normal. No test exercises a large sample counter.

Found by review; the same pattern was present in both trackers.

+++ src/rdf/north_ref_dpll.rs
@@ -624,8 +624,11 @@ impl DpllNorthTracker {
             // phase here would introduce reference drift across rotations.
             self.last_tick_sample = Some(compensated_sample);
             self.last_measured_sample = Some(compensated_sample);
-            self.last_tick_fraction =
-                fractional_sample_offset + (reported_sample as f32 - compensated_sample as f32);
+            // Differenced in integers: these are absolute sample counts, and
+            // past 2^24 an f32 cannot represent them to within a sample, so
+            // converting before subtracting loses the quantity being measured.
+            self.last_tick_fraction = fractional_sample_offset
+                + (reported_sample as i64 - compensated_sample as i64) as f32;
             ticks.push(NorthTick {
                 sample_index: reported_sample,
                 period: Some(period),
